requirements.txt
streamlit
fastapi
uvicorn
sqlalchemy
bcrypt
pyotp
qrcode
python-jose[cryptography]
passlib
requests
python-dotenv
apscheduler
pydantic-settings
